refactor: replace utils defaultBreakpoints with theme defaultBreakpoints
